+++ docs/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.16.5
+
+- 💥 [scrna.SeuratMetadataMutater] Move mutaters from in to envs
+- 🔥 [scrna.CellsDistribution] Remove unnecessary in.casefile
+- 🚑 [scrna.CellTypeAnnotation] Hotfix when envs.hitype_db as a file starts with "hitypedb_"
+
 ## 0.16.4
 
 - 🚑 [scrna.CellTypeAnnotation] Hotfix passing `envs.newcol`